+++ bundle/packages/vTeams/globals/globals.js
@@ -14,6 +14,12 @@ export const FORM_FIELDS = {
   MONTHLY_HOURS: 'Monthly Hours',
   ORGANIZATION: 'Organization',
   COMMENT: 'Comment',
+  TITLE: 'Title',
+  DATE_DUE: 'Requested Date Due',
+  DESCRIPTION: 'Description',
+  STATUS: 'Status',
+  BILLING_PERIOD: 'Billing Period',
+  ANNUAL_HOURS: 'Annual Hours',
 }
 
 export const NAMES = {