+++ bundle/packages/vTeams/src/redux/reducers.js
@@ -4,6 +4,8 @@ import { formResponses as values } from './modules/formResponses';
 import { tickets } from './modules/tickets';
 import { workLogs } from './modules/workLogs';
 import { organization } from './modules/organization';
+import { settings } from './modules/settings';
+import { clients } from './modules/clients';
 
 export default {
   app,
@@ -11,4 +13,6 @@ export default {
   tickets,
   workLogs,
   organization,
+  clients,
+  // settings,
 };